Unnecessary... explanation below
The problem is that when two blocked processes are meant to go to the
ready list at the same time, we have to check which one comes first in
the original list… that’s the one that will go first. Again, this is
only an issue if two blocked processes finish their IO at the same time.