Identified problem with picking next process to run (recall FCFS commits ago)
Bug: a process goes into the ready list when it’s unblocked or when its
quantum burst expires. Because of the order in which this is processed,
the blocked list will be processed before the running list, so if a
blocked element is freed it will come before the running element. This
is only a problem if it doesn’t follow the order of the text input.